requirements.txt
streamlit
torch
torchvision
pillow
scikit-learn
streamlit-drawable-canvas
numpy